Move file, change loopback handling, docs.
Move the ListeningPortsTest from android.net.cts to android.security.cts,
to better indicate that this is a security related test.

Start triggering test failures when a process is bound to the loopback
interface. In those cases, it's always better to use a UNIX domain
socket instead of a locally bound IP socket. UNIX domain sockets
can be protected with unix filesystem permissions, or
getsockopt(SO_PEERCRED) can be used to indicate who is on the
other end of the socket.

Drastically fix the documentation to explain why this test is present.

Modify the IP addresses and port numbers to be more human readable.

Change-Id: Id83aadd1455279a1a2d3fb4fb1915af558355aca