Refactored how data is shared between test apps.
When running the device-site tests, it's necessary to share state
between a second app and the main test app. Currently that's achieved
through a shared preference file that is accessed by both apps (since
they use a shared id), but this approach will not work on power save
mode tests (because the test app will be in foreground and the
background restrictions won't be applied in the second app).

This change refactors the data sharing mechanism by:

- Using an ordered broadcast that is sent from the test app to the
  secondary app.
- Checking for the network status in the secondary app.
- Moving the test logic to the client-side tests.

BUG:  27127112
Change-Id: I44987701b908b329fdf40e3a7a97e9f30cfadecb