Add assertion message to assertions for testExecute_withMobile.
Some partner test devices without a SIM Card or with no signal, thus
without mobile network data connection.

Added a message to check signal and SIM card and rerun the test.

Also added a verification that the network type is supported in the
event of a Connectivity Action's timeout. Unsuported network type could
be caused by a network stack crash.

Bug: 28207677
Change-Id: Ibca50359df7d5fb8371f4f4037fc81f029f48b30