Disable CtsNetTestCases in mainline-prod.
This currently still builds, but it will stop building as soon
as we make it depend on the networkstack current shims. It should
not be used for anything in mainline-prod anyway, everything uses
CtsNetTestCasesLatestSdk.

The Merged-In clause is from unrelated change aosp/1542627 and is
intended to ensure this change does not propagate to any
downstream branches.

Bug: 167645754
Test: mmm packages/modules/Connectivity
Change-Id: Ib47b181902e8e634f419e5544472859760c2ce77
Merged-In: Ib0ac49609e444a53a6fee4575f5078e15f364eef